#10 - Страница покупки. Завершение курса

+++ Shop/Data/Models/Car.cs
@@ -12,7 +12,7 @@ namespace Shop.Data.Models
         public string ShortDesc { get; set; }
         public string LongDesc { get; set; }
         public string Img { get; set; }
-        public ushort Price { get; set; }
+        public int Price { get; set; }
         public bool IsFavourite { get; set; }
         public bool Available { get; set; }
         public int CategoryId { get; set; }
